Refactor and modularize Javascript
